fix(recap): close review blockers on coordinates, races, and profiles
Coordinates: lat/lng are now clamped to their valid range and rounded
to 2 decimal places inside CreatePost itself, not trusted from the
client - the client-side rounding was advisory only, and a modified
client or a raw API call could otherwise store full-precision GPS
permanently. posts.lat/lng move from REAL to DOUBLE PRECISION so a
value that was already rounded round-trips exactly instead of picking
up a float32 representation error. Both are now exposed on Post
(coarse, same posture as the existing location string).

Manual recap duplication: CreateRecapPost's advisory lock and
in-transaction existence check previously only ran for the scheduled
origin, so two concurrent identical on-demand requests could both pass
the handler's cheap pre-check and both insert. The lock and re-check
now run for both origins, keyed on whatever makes each one unique
(cadence + period start for scheduled, period + panels for manual),
and the handler surfaces the authoritative conflict id from inside the
lock rather than only trusting its own pre-check.

Product decision: a recap post no longer appears on the authoring
admin's personal profile timeline (GET /api/users/{id}/posts) - it is
a group artifact, not something they personally posted. It stays
everywhere else. Feed gained an excludeRecap parameter for this.

On-demand custom periods longer than 366 days are now rejected with a
400 rather than accepted into a query that would scan the whole table.

+++ server/migrations/0018_recap.sql
@@ -52,6 +52,10 @@ ALTER TABLE server_config ADD COLUMN recap_since   TIMESTAMPTZ NOT NULL DEFAULT 
 -- Coordinates ship in v1 even though the map panel is v1.5, so two or three weeks of real
 -- data have accumulated by the time it needs them. The client rounds to 2 decimal places
 -- (~1.1km) before sending - strictly coarser than the "City, Country" string already
--- stored, so this leaks nothing new (see 0003's contract for that column).
-ALTER TABLE posts ADD COLUMN lat REAL;
-ALTER TABLE posts ADD COLUMN lng REAL;
+-- stored, so this leaks nothing new (see 0003's contract for that column). DOUBLE
+-- PRECISION, not REAL: the value is always rounded to 2 decimal places before it is
+-- written (normalizeCoord in queries.go), and REAL's float32 representation cannot hold
+-- that round trip exactly, which would make a stored value compare unequal to the one that
+-- was written.
+ALTER TABLE posts ADD COLUMN lat DOUBLE PRECISION;
+ALTER TABLE posts ADD COLUMN lng DOUBLE PRECISION;
